+++ lib/flutter_quill_delta_from_html.dart
@@ -1,7 +1,8 @@
 library flutter_quill_delta_from_html;
 
 export 'package:flutter_quill_delta_from_html/parser/html_to_delta.dart';
+export 'package:flutter_quill_delta_from_html/parser/colors.dart';
+export 'package:flutter_quill_delta_from_html/parser/custom_html_part.dart';
 export 'package:flutter_quill_delta_from_html/parser/html_utils.dart';
 export 'package:flutter_quill_delta_from_html/parser/html_to_operation.dart';
 export 'package:flutter_quill_delta_from_html/parser/extensions/node_ext.dart';
-