+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.1.5
+
+* Feat: added support for subscript and superscript
+* Feat: added support for color and background-color
+* Feat: added support for custom blocks 
+* Feat: added support for custom parsed `DOM Document` using `HtmlToDelta.convertDocument(DOMDocument)`
+* Chore: improved README 
+* Chore: improved documentation about project
+* Chore: now `resolveCurrentElement` was moved to the interface to give the same functionality to all implementations
+
 ## 1.1.0
 
 * Added support for fully attributes (could be partially bugged)